chore(e2e): drop the Go microbenchmark measurement path
Remove the Go microbenchmark ("bench") path from the e2e suite,
leaving the client-side characterization harness (the TestE2E_*
performance scenarios that emit workload-summary lines) as the single
performance measurement path. The bench path existed mainly to feed a
benchstat A/B comparator we have since decided not to build.

Deletes bench.go, the bench_test.go files, the bench Make targets, and
the Bench CI workflow; removes the bench scenario enumerators
(RunAllBenchmarks/RunBenchmarks/forEachBenchCell/benchmarkCases) and
the orphaned FD-limit and Azure shared-hyco-lease helpers that only
the benchmarks used. The shared dial helper is preserved (renamed
benchDial -> dialSender in performance.go) and the BackendScope unit
tests are relocated to scope_test.go.
